requirements.txt
discord.py[voice]
python-dotenv
alembic